Move GC specific types to gc module
