+++ rts/motoko-rts-tests/src/gc/heap.rs
@@ -1,7 +1,7 @@
 use super::utils::{write_word, ObjectIdx, GC, MAX_MARK_STACK_SIZE, WORD_SIZE};
 
+use motoko_rts::gc::mark_compact::mark_stack::INIT_STACK_SIZE;
 use motoko_rts::heap::Heap;
-use motoko_rts::mark_stack::INIT_STACK_SIZE;
 use motoko_rts::types::*;
 
 use std::cell::{Ref, RefCell};
