+++ rts/motoko-rts/src/gc/mark_compact/mark_stack.rs
